backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
psycopg[binary]==3.2.12
drf-spectacular==0.29.0
gunicorn==23.0.0
pillow==12.0.0
